Updating telemetry collection.
