+++ pkg/service-b/src/service.ts
@@ -44,26 +44,29 @@ export function createServiceB(getMetaData: () => unknown): ServiceB {
       );
     },
     chainForward: async (result) => {
-      return withSpanAsync('chainForward', () =>
-        getRemoteServiceC().chainForward({
+      return withSpanAsync('chainForward', async () => {
+        const service = await getRemoteServiceC();
+        return await service.chainForward({
           ...result,
           b: createResult(),
           order: [...result.order, 'b'],
-        })
-      );
+        });
+      });
     },
     chainBackward: async (result) => {
-      return withSpanAsync('chainBackward', () =>
-        getRemoteServiceA().chainBackward({
+      return withSpanAsync('chainBackward', async () => {
+        const service = await getRemoteServiceA();
+        return await service.chainBackward({
           ...result,
           b: createResult(),
           order: [...result.order, 'b'],
-        })
-      );
+        });
+      });
     },
     transformFromA: async () => {
       return withSpanAsync('transformFromA', async () => {
-        const result = await getRemoteServiceA().doSomething();
+        const service = await getRemoteServiceA();
+        const result = await service.doSomething();
         return logData({
           fromA: result,
           message: 'Transformed by Service B',
@@ -72,7 +75,8 @@ export function createServiceB(getMetaData: () => unknown): ServiceB {
     },
     transformFromC: async () => {
       return withSpanAsync('transformFromC', async () => {
-        const result = await getRemoteServiceC().doSomething();
+        const service = await getRemoteServiceC();
+        const result = await service.doSomething();
         return logData({
           fromC: result,
           message: 'Transformed by Service B',
@@ -81,7 +85,8 @@ export function createServiceB(getMetaData: () => unknown): ServiceB {
     },
     transformFromD: async () => {
       return withSpanAsync('transformFromD', async () => {
-        const result = await getRemoteServiceD().doSomething();
+        const service = await getRemoteServiceD();
+        const result = await service.doSomething();
         return logData({
           fromD: result,
           message: 'Transformed by Service B',
